# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GuardianWAF API
  description: GuardianWAF REST API for security monitoring, configuration, and management.
  version: 1.0.0
  contact:
    name: GuardianWAF
    url: https://github.com/guardianwaf/guardianwaf
  license:
    name: MIT
servers:
  - url: https://localhost:9443
    description: Default dashboard endpoint
security:
  - ApiKeyAuth: []
  - SessionAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: gwaf_session
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
paths:
...
  /api/admin/tenants:
    get:
      tags: [Tenants]
      summary: List all tenants
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Tenant list
    post:
      tags: [Tenants]
      summary: Create a tenant
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                domains:
                  type: array
                  items:
                    type: string
                quota:
                  type: object
      responses:
        '200':
          description: Tenant created
  /api/admin/tenants/{id}:
    get:
      tags: [Tenants]
      summary: Get tenant details
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant details
    put:
      tags: [Tenants]
      summary: Update tenant
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Tenant updated
    delete:
      tags: [Tenants]
      summary: Delete tenant
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant deleted
...
  /api/admin/stats:
    get:
      tags: [Tenants]
      summary: Get platform statistics
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Platform statistics
  /api/admin/billing:
    get:
      tags: [Tenants]
      summary: List all invoices
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Invoice list
...
  /api/admin/alerts:
    get:
      tags: [Tenants]
      summary: Get cross-tenant alerts
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Alert list
  /api/admin/usage:
    get:
      tags: [Tenants]
      summary: Get all tenant usage
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Usage data
...